Add wifiloader
This is needed on some Samsung device to load the wifi module. Before
this was solved using a bash script which does:

  cat /proc/deferred_initcalls

but the new sepolicy rules don't allow init execute bash scripts. For
this we need a new tool which does the job and we can create sepolicy
rule for it.

Change-Id: I137cfaaff74955ad431bc09d74b0c970618dc3bf
Signed-off-by: Andreas Schneider <asn@cryptomilk.org>